+++ src/enums/remote-branch-global-commands.enum.ts
@@ -1,0 +1,3 @@
+export enum RemoteBranchGlobalCommands {
+    refresh = 'scm-remote-branches.refresh'
+}