+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { BranchSwitcher } from './branches';
 import { StashManager } from './stashes';
+import { RemoteBranchManager } from './remote-branches';
 import { Git } from './git';
 
 export function activate(context: vscode.ExtensionContext) {
@@ -9,4 +10,5 @@ export function activate(context: vscode.ExtensionContext) {
 
     new BranchSwitcher(context, git);
     new StashManager(context, git);
+    new RemoteBranchManager(context, git);
 }