+++ src/models/remote-branch.ts
@@ -1,0 +1,8 @@
+import { Repository } from "../typings/git-extension";
+
+export interface RemoteBranch {
+    readonly repo: Repository;
+    readonly remote?: string;
+    readonly branchName?: string;
+    readonly commit?: string;
+};